Allow HTML in the description of the image section
We have a rich text editor for the description of the image field,
but we are stripping HTML so styling is not displayed. This PR removes
HTML stripping.

Additionally truncation is also removed as it can break HTML tags and
cause issues in the layout.

Example:
`<p>Some long text</p>` can get truncated to `<p>Some long` which
introduces an unclosed tag and breaks the layout.